% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pediarr.R
\name{pediacategory}
\alias{pediacategory}
\title{List out members of Wikipedia category}
\usage{
pediacategory(wikicategory, lang = "en", limit = 500,
  cmtype = "page|subcat|file")
}
\arguments{
\item{wikicategory}{A character string of Wikipedia category}

\item{lang}{Language of Wikipedia to query, default = 'en'}

\item{limit}{integer, number of items to return}

\item{cmtype}{A character string represents member type to be list out, default is pages, subcategories and files}
}
\value{
vector of wikipedia titles
}
\description{
List out members of Wikipedia category
}
\examples{
pediacategory("Category:Dynamically typed programming languages")
pediacategory("Category:Dynamically typed programming languages", cmtype = 'page') # ignore subcat and files
}

