% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pediarr.R
\name{pediaextract}
\alias{pediaextract}
\title{Retrieve the extract of Wikipedia article}
\usage{
pediaextract(wikititles, lang = "en")
}
\arguments{
\item{wikititles}{A character string or Charactor vector of Wikipedia article titles}

\item{lang}{Language of Wikipedia to query, default = 'en'}
}
\value{
data.frame of extracts
}
\description{
Retrieve the first section (extract) of Wikipedia articles
}
\examples{
pediaextract("Hong_Kong")
pediaextract("Hong Kong") # smart enough to replace space with underscore
pediaextract("Hong_Kong", lang="es")
pediaextract(c("R.E.M.", "Nirvana (band)", "Pearl Jam"))
}

