% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pediarr.R
\name{pediafulltext}
\alias{pediafulltext}
\title{Retrieve the full text of Wikipedia article}
\usage{
pediafulltext(wikititle, lang = "en", format = "text")
}
\arguments{
\item{wikititle}{Character string of Wikipedia article title}

\item{lang}{language of Wikipedia to query, default = 'en'}

\item{format}{character string of either 'text', 'wikimarkup' or 'html', default 'text'}
}
\value{
Character string of Wikipedia full text
}
\description{
Retrieve the full text of Wikipedia article
}
\examples{
pediafulltext("Albert Einstein") # very messy
pediafulltext("Hong Kong", format = 'html')
pediafulltext("Hong Kong", lang="es", format = 'wikimarkup')
}

