% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pediarr.R
\name{pedialang}
\alias{pedialang}
\title{Retrieve the wikipedia title in other language}
\usage{
pedialang(wikititle, lang = "en", limit = 50, lllang = NULL)
}
\arguments{
\item{wikititle}{A character string of Wikipedia title to query}

\item{lang}{Language of Wikipedia to query, default = 'en'}

\item{limit}{integer, number of items to return}

\item{lllang}{A character string represents the language of returned title. If NULL, all languages will be returned}
}
\value{
data.frame of wikipedia titles
}
\description{
Retrieve the wikipedia title in other language
}
\examples{
pedialang("Albert Einstein")
pedialang("John_McCarthy_(computer_scientist)", lllang = "zh")
pedialang("Alanu Turing", lang = 'co', lllang = 'ko')
}

