% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pediarr.R
\name{pediasearch}
\alias{pediasearch}
\title{Search for Wikipedia articles}
\usage{
pediasearch(searchstring, lang = "en", extract = FALSE, limit = 20,
  namespace = 0)
}
\arguments{
\item{searchstring}{A character string of search string}

\item{lang}{Language of Wikipedia to query, default = 'en'}

\item{extract}{Boolean, to retrieve the extracts or not, only available when namespace = 0}

\item{limit}{integer, number of items to return}

\item{namespace}{integer, namespace to search for, common namespace: 0 = articles, 14 = categories}
}
\value{
vector of wikipedia titles or vector of extracts
}
\description{
Retrieve the titles of Wikipedia items based on search string
}
\examples{
pediasearch("ukulele")
pediasearch("ukulele", extract = TRUE, limit = 1)
pediasearch("Python")
pediasearch("Python", namespace = 14) # search for categories
pediasearch("Dynamically typed programming languages", namespace = 14, limit = 1)
}

